front-end service catalog moved to models-library
